New schema to avoid call systems. It consists of a modified version of OLAM_filelist that receives a flag (nocall) to decide to run the system subroutine or not.
If nocall is .FALSE., then the old way (using system) will run.

If nocall is .TRUE., then the new way will run. In this case the variable (IAPR, SST_DATABASE or SEAICE_DATABASE) on namelist will be used as the list of files, in the same way that tmpname inside filelist is on the original version.

Summary of changes:
- hdf5_utils.f90, olammain.F90: removal of (unnecessary?) "rm -f"
- filelist.f90, seaice_database_read.f90, sst_database_read.f90, file_inv.f90: implementation of the new schema
- namelists and oname_check.f90: documentation and new tests on the changed namelists flags



git-svn-id: https://svn.code.sf.net/p/olam-model/code/branches/pedropaislopes_si@18 2bc74c4e-f164-4307-89b2-97abcff1156f